+++ Models/Responses/Nurse.cs
@@ -10,7 +10,7 @@ namespace dotnet.Models
 
         public int DutyDuration { get; set; }
         public int SharePercentage { get; set; }
-        public double Salary { get; set; }
+        public int Salary { get; set; }
 
         public virtual User User { get; set; }
